+++ app/models/Baseline.scala
@@ -3,13 +3,13 @@ package models
 import play.api.db.slick.Config.driver.simple._
 import java.sql.Date
 
-case class Baseline(id: Int, name: String, year: Date, description: String)
+case class Baseline(id: Long, name: String, year: Date, description: String)
 
 class BaselineTable(tag: Tag) extends Table[Baseline](tag, "BASELINE") {
   def * = (id, name, year, description) <> (Baseline.tupled, Baseline.unapply)
   def ? = (id.?, name.?, year.?, description.?).shaped.<>({ r => import r._; _1.map(_ => Baseline.tupled((_1.get, _2.get, _3.get, _4.get))) }, (_: Any) => throw new Exception("Inserting into ? projection not supported."))
 
-  val id: Column[Int] = column[Int]("ID", O.AutoInc, O.PrimaryKey)
+  val id: Column[Long] = column[Long]("ID", O.AutoInc, O.PrimaryKey)
   val name: Column[String] = column[String]("NAME")
   val year: Column[Date] = column[Date]("YEAR")
   val description: Column[String] = column[String]("DESCRIPTION")
@@ -17,7 +17,7 @@ class BaselineTable(tag: Tag) extends Table[Baseline](tag, "BASELINE") {
 
 object Baselines extends DAO {
 
-  def findById(id: Int)(implicit session: Session): Option[Baseline] =
+  def findById(id: Long)(implicit session: Session): Option[Baseline] =
     Baselines.filter(_.id === id).firstOption
 
   def findByName(name: String)(implicit session: Session): Option[Baseline] =
@@ -36,7 +36,7 @@ object Baselines extends DAO {
   def all(implicit session: Session): List[Baseline] =
     Baselines.list
 
-  def insert(a: Baseline)(implicit session: Session): Int =
+  def insert(a: Baseline)(implicit session: Session): Long =
     (Baselines returning Baselines.map(_.id)) += a
 
 }
